requirements.txt
Django==3.2.5
asgiref==3.4.1
djangorestframework==3.12.4
pip==21.1.3
psycopg2-binary==2.9.1
pytz==2021.1
setuptools==57.0.0
sqlparse==0.4.1